RAG_agent/requirements.txt
# Core dependencies
langchain>=1.0.8
langchain-community>=0.3.0
langgraph>=1.0.3
chromadb>=1.3.5
langchain-google-genai>=3.1.0
langchain-chroma>=1.0.0
pydantic>=2.12.4
python-dotenv>=1.2.1
rank_bm25>=0.2.2
rapidfuzz>=3.0.0
tqdm>=4.67.1
langchain-text-splitters>=1.0.0
httpx>=0.25.0
pyyaml>=6.0.0
html2text>=2020.1.16

# Security validation (ML-based prompt injection detection)
transformers>=4.36.0
torch>=2.1.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Web search (online mode)
tavily-python>=0.3.0

# Embeddings (Free local embeddings via Hugging Face)
sentence-transformers>=5.1.0
langchain-huggingface>=1.1.0

# Observability and debugging
langsmith>=0.1.0
